split up tests into text and img tests
