+++ package.json
@@ -11,6 +11,10 @@
 		"dist": "cd ./test; gulp dist --env dist",
 		"pretest": "npm run babel-dist; npm run clean; npm run watchtest; cd ./test; gulp test --env dist",
 		"test": "jest",
+		"pretest-txt": "npm run babel-dist; npm run clean; npm run watchtest; cd ./test; gulp testTxt --env dist",
+		"test-txt": "npm run jest --runTestsByPath txt.test.js",
+		"pretest-img": "cd ./test; gulp testImg --env dist",
+		"test-img": "npm run jest --runTestsByPath img.test.js",
 		"jest": "jest",
 		"clean": "rm -rf ./test/output",
 		"testd": "http-server ./test -o",
